upnext

+++ main.js
@@ -1,12 +1,9 @@
 const discord = require('discord.js'); //up
-const {Client, MessageAttachment, MessageEmbed} = require('discord.js');
+const {Client} = require('discord.js');
 const bot = new Client();
 const client = new discord.Client();
 const prefix = '/';
 const fs = require('fs');
-const cheerio = require('cheerio');
-const request = require('request');
-
 
 client.commands = new discord.Collection();
 
@@ -28,27 +25,43 @@ client.on('message', async message => {
 
     if(command === "hei")
     {
-        if(message.member.voice.channel) {
+        if(message.member.voice.channel && message.member.roles.cache.has('769836747533713438')) 
+        {
             message.delete();
             let replies = ['heyhey.mp3', 'itzyhey.mp3']
             let random = Math.floor(Math.random() * replies.length);
             const connection = await message.member.voice.channel.join();
             connection.play(replies[random]);
-            setTimeout(() => {
+            setTimeout(() => 
+            {
                 connection.disconnect();
             }, 3000)
         }
-    }   
+    }  
     else if(command === "win")
     {
-        if(message.member.voice.channel) {
+        if(message.member.voice.channel && message.member.roles.cache.has('769836747533713438')) 
+        {
             message.delete();
             const connec = await message.member.voice.channel.join();
             connec.play('ez4ence.mp3');
-            setTimeout(() => {
+            setTimeout(() => 
+            {
                 connec.disconnect();
             }, 6500)
         }
+    }
+    else if(command === 'play')
+    {
+        client.commands.get('play').execute(message, args);
+    }
+    else if(command === 'bye')
+    {
+        client.commands.get('bye').execute(message, args);
+    }
+    else if(command === 'clear')
+    {
+        client.commands.get('clear').execute(message, args);
     }
     else if(command === 'ping'){
         client.commands.get('ping').execute(message, args);
@@ -119,28 +132,7 @@ client.on('message', async message => {
     }
     else if(command === `show`)
     {
-        var options = {
-            url: "http://results.dogpile.com/serp?qc=images&q=" + args,
-            method: "GET",
-            headers: {
-                "Accept": "text/html",
-                "User-Agent": "Chrome"
-            }
-        };
-        request(options, function(error, response, responseBody) {
-            if (error) {
-                return;
-            }
-
-            $ = cheerio.load(responseBody);
-            var links = $(".image a.link");
-            var urls = new Array(links.length).fill(0).map((v, i) => links.eq(i).attr("href"));
-            console.log(urls);
-            if (!urls.length) {   
-                return;
-            }
-            message.channel.send( urls[Math.floor(Math.random() * urls.length)]);
-        });
+        client.commands.get('show').execute(message, args);
     }
 });
